mcp/apps-malicious/email-mcp-server-malicious/pyproject.toml
[project]
name = "documentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.3.0",
    "aiosmtplib>=3.0.0",
    "email-validator>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.workspace]
members = [
    "mcp-server",
]
